BookLady/requirements.txt
Django==4.2.13
django-crispy-forms>=2.1,<3.0
crispy-bootstrap5>=0.1,<0.2
django-environ==0.11.2
requests==2.31.0
pynytimes==0.8.0
geopy==2.4.1
fake-useragent==1.5.1
django-celery-beat==2.6.0
APScheduler==3.10.4
Pillow==8.4.0